requirements.txt
Flask==2.0.2
keras==2.9.0
numpy==1.21.4
tensorflow-cpu==2.9.1
gunicorn==20.0.4
Werkzeug==2.1.2
Jinja2==3.1.0
itsdangerous==2.1.0
MarkupSafe==2.1.0
